Move PlanSimulation to foundation
