+++ src/main/kotlin/gov/nasa/jpl/pyre/examples/lander/activities/apss/APSSGeneric.kt
@@ -8,7 +8,7 @@ import gov.nasa.jpl.pyre.kernel.times
 import gov.nasa.jpl.pyre.examples.lander.Mission
 import gov.nasa.jpl.pyre.examples.lander.models.data.DataConfig.APID.*
 import gov.nasa.jpl.pyre.examples.lander.models.power.PowerModel
-import gov.nasa.jpl.pyre.general.plans.Activity
+import gov.nasa.jpl.pyre.foundation.plans.Activity
 import gov.nasa.jpl.pyre.foundation.resources.discrete.DiscreteResourceOperations.set
 import gov.nasa.jpl.pyre.foundation.tasks.TaskOperations.delay
 import gov.nasa.jpl.pyre.foundation.tasks.TaskScope
