+++ src/main/kotlin/gov/nasa/jpl/pyre/examples/lander/activities/apss/APSSPaeDataRecovery.kt
@@ -7,7 +7,7 @@ import gov.nasa.jpl.pyre.kernel.ratioOver
 import gov.nasa.jpl.pyre.kernel.times
 import gov.nasa.jpl.pyre.examples.lander.Mission
 import gov.nasa.jpl.pyre.examples.lander.models.data.DataConfig.APID.APID_APSS_CONTINUOUS_SCI
-import gov.nasa.jpl.pyre.general.plans.Activity
+import gov.nasa.jpl.pyre.foundation.plans.Activity
 import gov.nasa.jpl.pyre.foundation.tasks.TaskScope
 import kotlinx.serialization.Serializable
