+++ src/main/kotlin/gov/nasa/jpl/pyre/examples/orbiter/power/ChangePowerDemand.kt
@@ -1,6 +1,6 @@
 package gov.nasa.jpl.pyre.examples.orbiter.power
 
-import gov.nasa.jpl.pyre.general.plans.Activity
+import gov.nasa.jpl.pyre.foundation.plans.Activity
 import gov.nasa.jpl.pyre.foundation.resources.discrete.DoubleResourceOperations.increase
 import gov.nasa.jpl.pyre.foundation.tasks.TaskScope
 import kotlinx.serialization.Serializable
