+++ src/main/kotlin/gov/nasa/jpl/pyre/examples/scheduling/gnc/activities/GncSetAgility.kt
@@ -2,7 +2,7 @@ package gov.nasa.jpl.pyre.examples.scheduling.gnc.activities
 
 import gov.nasa.jpl.pyre.examples.scheduling.gnc.model.GncModel
 import gov.nasa.jpl.pyre.examples.scheduling.gnc.model.MRAD_PER_SECOND
-import gov.nasa.jpl.pyre.general.plans.Activity
+import gov.nasa.jpl.pyre.foundation.plans.Activity
 import gov.nasa.jpl.pyre.general.resources.discrete.unit_aware.QuantityResourceOperations.set
 import gov.nasa.jpl.pyre.general.units.UnitAware.Companion.times
 import gov.nasa.jpl.pyre.foundation.tasks.TaskScope
