+++ src/main/kotlin/gov/nasa/jpl/pyre/examples/scheduling/imager/activities/ImagerDoObservation.kt
@@ -3,8 +3,8 @@ package gov.nasa.jpl.pyre.examples.scheduling.imager.activities
 import gov.nasa.jpl.pyre.kernel.Duration
 import gov.nasa.jpl.pyre.examples.scheduling.imager.model.ImagerModel
 import gov.nasa.jpl.pyre.examples.scheduling.imager.model.ImagerModel.ImagerMode
-import gov.nasa.jpl.pyre.general.plans.Activity
-import gov.nasa.jpl.pyre.general.plans.ActivityActions.spawn
+import gov.nasa.jpl.pyre.foundation.plans.Activity
+import gov.nasa.jpl.pyre.foundation.plans.ActivityActions.spawn
 import gov.nasa.jpl.pyre.foundation.resources.discrete.DiscreteResourceOperations.equals
 import gov.nasa.jpl.pyre.foundation.resources.discrete.DiscreteResourceOperations.set
 import gov.nasa.jpl.pyre.foundation.resources.getValue
