+++ src/main/kotlin/gov/nasa/jpl/pyre/examples/scheduling/system/activities/GncActivity.kt
@@ -3,7 +3,7 @@ package gov.nasa.jpl.pyre.examples.scheduling.system.activities
 
 import gov.nasa.jpl.pyre.examples.scheduling.gnc.model.GncModel
 import gov.nasa.jpl.pyre.examples.scheduling.system.model.SystemModel
-import gov.nasa.jpl.pyre.general.plans.Activity
+import gov.nasa.jpl.pyre.foundation.plans.Activity
 import gov.nasa.jpl.pyre.foundation.tasks.TaskScope
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
