+++ src/main/kotlin/gov/nasa/jpl/pyre/examples/sequencing/commands/seq/SeqIf.kt
@@ -2,7 +2,7 @@ package gov.nasa.jpl.pyre.examples.sequencing.commands.seq
 
 import gov.nasa.jpl.pyre.examples.sequencing.SequencingDemo
 import gov.nasa.jpl.pyre.examples.sequencing.fsw.FswModel.GlobalIntVarName
-import gov.nasa.jpl.pyre.general.plans.Activity
+import gov.nasa.jpl.pyre.foundation.plans.Activity
 import gov.nasa.jpl.pyre.foundation.tasks.TaskScope
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
