+++ src/main/kotlin/gov/nasa/jpl/pyre/examples/sequencing/commands/seq/SeqSetGlobalInt.kt
@@ -3,7 +3,7 @@ package gov.nasa.jpl.pyre.examples.sequencing.commands.seq
 import gov.nasa.jpl.pyre.examples.sequencing.SequencingDemo
 import gov.nasa.jpl.pyre.examples.sequencing.fsw.FswModel.GlobalIntVarName
 import gov.nasa.jpl.pyre.examples.sequencing.primeness.SideIndicator
-import gov.nasa.jpl.pyre.general.plans.Activity
+import gov.nasa.jpl.pyre.foundation.plans.Activity
 import gov.nasa.jpl.pyre.foundation.resources.discrete.DiscreteResourceOperations.set
 import gov.nasa.jpl.pyre.foundation.tasks.TaskScope
 import kotlinx.serialization.SerialName
