+++ src/main/kotlin/gov/nasa/jpl/pyre/examples/sequencing/commands/telecom/SetPrimeRadio.kt
@@ -2,7 +2,7 @@ package gov.nasa.jpl.pyre.examples.sequencing.commands.telecom
 
 import gov.nasa.jpl.pyre.examples.sequencing.SequencingDemo
 import gov.nasa.jpl.pyre.examples.sequencing.primeness.Side
-import gov.nasa.jpl.pyre.general.plans.Activity
+import gov.nasa.jpl.pyre.foundation.plans.Activity
 import gov.nasa.jpl.pyre.foundation.resources.discrete.DiscreteResourceOperations.set
 import gov.nasa.jpl.pyre.foundation.tasks.TaskScope
 import kotlinx.serialization.SerialName
