+++ src/main/kotlin/gov/nasa/jpl/pyre/foundation/plans/InstantSerializer.kt
@@ -1,4 +1,4 @@
-package gov.nasa.jpl.pyre.general.plans
+package gov.nasa.jpl.pyre.foundation.plans
 
 import gov.nasa.jpl.pyre.utilities.InvertibleFunction
 import gov.nasa.jpl.pyre.kernel.Serialization.alias
@@ -6,8 +6,8 @@ import kotlinx.serialization.KSerializer
 import kotlinx.serialization.builtins.serializer
 import kotlin.time.Instant
 
-class InstantSerializer : KSerializer<Instant> by String.Companion.serializer().alias(
-    InvertibleFunction.Companion.of(
+class InstantSerializer : KSerializer<Instant> by String.serializer().alias(
+    InvertibleFunction.of(
     Instant.Companion::parse,
     Instant::toString
 ))
