+++ src/main/kotlin/gov/nasa/jpl/pyre/general/scheduling/SchedulingSystem.kt
@@ -4,11 +4,11 @@ import gov.nasa.jpl.pyre.utilities.andThen;
 import gov.nasa.jpl.pyre.kernel.InconProvider
 import gov.nasa.jpl.pyre.kernel.JsonConditions
 import gov.nasa.jpl.pyre.kernel.ReportHandler
-import gov.nasa.jpl.pyre.general.plans.Activity
-import gov.nasa.jpl.pyre.general.plans.ActivityActions.ActivityEvent
-import gov.nasa.jpl.pyre.general.plans.GroundedActivity
-import gov.nasa.jpl.pyre.general.plans.Plan
-import gov.nasa.jpl.pyre.general.plans.PlanSimulation
+import gov.nasa.jpl.pyre.foundation.plans.Activity
+import gov.nasa.jpl.pyre.foundation.plans.ActivityActions.ActivityEvent
+import gov.nasa.jpl.pyre.foundation.plans.GroundedActivity
+import gov.nasa.jpl.pyre.foundation.plans.Plan
+import gov.nasa.jpl.pyre.foundation.plans.PlanSimulation
 import gov.nasa.jpl.pyre.general.reporting.ReportHandling.assumeType
 import gov.nasa.jpl.pyre.general.reporting.ReportHandling.channels
 import gov.nasa.jpl.pyre.general.results.SimulationResults
