+++ src/main/kotlin/gov/nasa/jpl/pyre/general/testing/SimulationResults.kt
@@ -1,7 +1,7 @@
 package gov.nasa.jpl.pyre.general.testing
 
-import gov.nasa.jpl.pyre.general.plans.Activity
-import gov.nasa.jpl.pyre.general.plans.ActivityActions.ActivityEvent
+import gov.nasa.jpl.pyre.foundation.plans.Activity
+import gov.nasa.jpl.pyre.foundation.plans.ActivityActions.ActivityEvent
 import gov.nasa.jpl.pyre.foundation.reporting.ChannelizedReport
 import kotlin.time.Instant
